move mongo_client module to lib
